move serializing logic to serializers file
